feat(ai-label): add typescript types
